Delete bbcode and lower reverse mode from IRCode
