+++ src/Mooncake.jl
@@ -179,9 +179,6 @@ include("debug_mode.jl")
 include("stack.jl")
 
 @unstable begin
-include(joinpath("interpreter", "bbcode.jl"))
-using .BasicBlockCode
-
 include(joinpath("interpreter", "contexts.jl"))
 include(joinpath("interpreter", "abstract_interpretation.jl"))
 include(joinpath("interpreter", "patch_for_319.jl"))
